requirements.txt
selenium=3.14.1
bs4=0.0.1
pillow=5.4.1